Move Netty implementation classes to a separate package
