+++ src/main/kotlin/io/libp2p/transport/implementation/P2PChannelOverNetty.kt
@@ -1,4 +1,4 @@
-package io.libp2p.transport.tcp
+package io.libp2p.transport.implementation
 
 import io.libp2p.etc.types.toVoidCompletableFuture
 import io.netty.channel.Channel
