+++ src/test/kotlin/io/libp2p/tools/TestChannel.kt
@@ -1,7 +1,7 @@
 package io.libp2p.tools
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder
-import io.libp2p.transport.tcp.ConnectionOverNetty
+import io.libp2p.transport.implementation.ConnectionOverNetty
 import io.libp2p.etc.CONNECTION
 import io.libp2p.etc.types.lazyVar
 import io.libp2p.etc.util.netty.nettyInitializer
@@ -25,7 +25,13 @@ class TestChannel(id: String = "test", initiator: Boolean, vararg handlers: Chan
     EmbeddedChannel(
         TestChannelId(id),
         nettyInitializer {
-            it.attr(CONNECTION).set(ConnectionOverNetty(it, NullTransport(), initiator))
+            it.attr(CONNECTION).set(
+                ConnectionOverNetty(
+                    it,
+                    NullTransport(),
+                    initiator
+                )
+            )
         },
         *handlers
     ) {
